Upgraded gems, rspec configs, and split files
