+++ spec/methods.rb
